Run test foked
