+++ dd-java-agent/instrumentation/kafka-clients-3.8/src/test/groovy/KafkaReactorForkedTest.groovy
@@ -26,7 +26,7 @@ import java.util.concurrent.CountDownLatch
 import java.util.concurrent.LinkedBlockingQueue
 import java.util.concurrent.TimeUnit
 
-class KafkaReactorTest extends AgentTestRunner {
+class KafkaReactorForkedTest extends AgentTestRunner {
 
   final ConfluentKafkaContainer kafkaContainer = new ConfluentKafkaContainer(DockerImageName.parse("confluentinc/cp-kafka:7.7.1"))
